pyproject.toml
[tool.pyright]
venvPath = "."
venv = ".venv"

[project]
name = "minigrim0"
version = "0.1.0"
description = "Minigrim0's website"
requires-python = ">=3.11"
authors = [{name = "Minigrim0", email="grimauflorent@gmail.com"}]
license = "MIT"
readme = "README.md"
dependencies = [
    "coloredlogs == 15.0.1",
    "django == 5.1.1",
    "django-autocomplete-light>=3.11.0",
    "django-colorfield == 0.11.0",
    "django-environ == 0.11.2",
    "django-import-export>=4.3.1",
    "faker == 28.4.1",
    "mistune == 3.0.2",
    "psycopg2 == 2.9.9",
    "pygments == 2.18.0",
    "redis == 5.1.0",
]

[dependency-groups]
dev = ["isort == 5.13.2", "pylint == 2.17.7", "black == 23.12.1"]
prod = ["gunicorn == 21.2.0"]

[tool.black]
line-length = 110
target-version = ["py312"]
extend-exclude = '''
.*\/migrations\/.*\.py  # Skip all migration files
\.venv\/.*  # Skip virtual environment
'''

[tool.pylint]
ignore = [
    "migrations",
    ".venv",
]
disable = [
    "too-few-public-methods",
]
jobs = 4
